+++ src/EasyConsoleApplication/Application.cs
@@ -11,7 +11,7 @@ namespace EasyConsoleApplication
         public static void Render(Menu menu)
         {
             MenuRendering = new Rendering();
-            MenuRendering.Render(menu.Title, menu.Color, null, ConsoleColor.White, menu);
+            MenuRendering.Render(menu.Title, menu.Color, null, ConsoleSettings.DefaultColor, menu);
         }
 
         private static readonly Router Router = new Router();