+++ src/EasyConsoleApplication/Menus/Menu.cs
@@ -6,13 +6,18 @@ namespace EasyConsoleApplication.Menus
     public class Menu : IMenuItem
     {
         public string Title { get; set; }
-        public ConsoleColor Color { get; set; } = ConsoleColor.White;
+        public ConsoleColor Color { get; set; } = ConsoleSettings.DefaultColor;
 
         public List<IMenuItem> Items { get; set; } = new List<IMenuItem>();
 
         internal Menu() { }
 
-        public Menu(string title, ConsoleColor color = ConsoleColor.White)
+        public Menu(string title)
+        {
+            Title = title;
+        }
+
+        public Menu(string title, ConsoleColor color)
         {
             Title = title;
             Color = color;