+++ src/EasyConsoleApplication/Menus/MenuItem.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 
 namespace EasyConsoleApplication.Menus
 {
@@ -11,9 +12,10 @@ namespace EasyConsoleApplication.Menus
         public string Command { get; private set; }
 
         public string Title { get; private set; }
-        public ConsoleColor Color { get; set; } = ConsoleColor.White;
+        public ConsoleColor Color { get; set; } = ConsoleSettings.DefaultColor;
 
         public Action Action { get; private set; }
+        public Func<Task> ActionAsync { get; private set; }
 
         public MenuItem(string title, Action action)
             : this(null, title, action)
@@ -24,6 +26,18 @@ namespace EasyConsoleApplication.Menus
         {
             Title = title;
             Action = action;
+            Command = command;
+        }
+
+        public MenuItem(string title, Func<Task> actionAsync)
+            : this(null, title, actionAsync)
+        {
+        }
+
+        public MenuItem(string command, string title, Func<Task> actionAsync)
+        {
+            Title = title;
+            ActionAsync = actionAsync;
             Command = command;
         }
     }