+++ Changelog.md
@@ -1,3 +1,9 @@
+## 0.4.0
+
+- ConsoleSettings: allows to set some global settings.
+- Async function support for MenuItems.
+- Fixed: Out of Bounds bug when selecting menu items.
+
 ## 0.3.0
 
-- added net5.0 support
+- Added net5.0 support.